requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
boto3==1.38.39
botocore==1.38.39
certifi==2025.4.26
cffi==1.16.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.9.1
crontab==1.0.4
cryptography==42.0.5
decorator==4.4.2
diff-match-patch==20241021
Django==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-import-export==4.3.8
django-redis==5.4.0
django-rq==3.0.1
django-storages==1.14.6
djangorestframework==3.16.0
freezegun==1.5.2
gunicorn==23.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
jmespath==1.0.1
moviepy==1.0.3
numpy==2.1.2
packaging==25.0
paramiko==3.4.0
pillow==11.2.1
pluggy==1.6.0
proglog==0.1.12
psycopg2-binary==2.9.10
pycparser==2.21
Pygments==2.19.1
PyNaCl==1.5.0
pytest==8.4.0
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
rq==2.3.3
rq-scheduler==0.14.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
